requirements.txt
pytest
opencv-python-headless>=4.7.0
Pillow>=9.0
